Exception Hadling and SOP
The ActionResult class is a custom class designed to encapsulate the result and status of an operation, typically used in automated testing or exception handling. It is a utility class that holds two pieces of information:

Message: This typically contains a descriptive message, such as a log or status update, that explains what happened during the execution of an action.
Status: This indicates the outcome of the action (e.g., success, fail) or other status indicators that define the result of the operation.